+++ src/enums/schemasMap.ts
@@ -23,9 +23,9 @@ export default {
   [services.authRoleAssignment]: 'azureAuthRoleAssignment',
   [services.authRoleDefinition]: 'azureAuthRoleDefinition',
   [services.autoProvisioningSettings]: 'azureAutoProvisioningSetting',
-  [services.backupVaults]: 'azureBackupVaults',
-  [services.backupInstances]: 'azureBackupInstances',
-  [services.backupPolicies]: 'azureBackupPolicies',
+  [services.backupVaults]: 'azureBackupVault',
+  [services.backupInstances]: 'azureBackupInstance',
+  [services.backupPolicies]: 'azureBackupPolicy',
   [services.billing]: 'azureBilling',
   [services.cdnCustomDomains]: 'azureCdnCustomDomain',
   [services.cdnEndpoints]: 'azureCdnEndpoint',
@@ -67,8 +67,8 @@ export default {
   [services.privateDns]: 'azurePrivateDnsZone',
   [services.publicIp]: 'azurePublicIp',
   [services.recoveryVaults]: 'azureRecoveryVault',
-  [services.recoveryInstances]: 'azureRecoveryInstances',
-  [services.recoveryPolicies]: 'azureRecoveryPolicies',
+  [services.recoveryInstances]: 'azureRecoveryInstance',
+  [services.recoveryPolicies]: 'azureRecoveryPolicy',
   [services.redisCaches]: 'azureRedisCache',
   [services.replicationAppliances]: 'azureReplicationAppliance',
   [services.replicationCenters]: 'azureReplicationCenter',
