+++ src/services/backupPolicies/format.ts
@@ -1,13 +1,13 @@
 import cuid from 'cuid'
 import { isEmpty } from 'lodash'
+import { BackupPolicy, TriggerContext } from '../../types'
 import {
   AzureBackupPolicy,
   AzureBackupPolicyProperties,
-  AzureBackupPolicyTriggerContext
+  AzureBackupPolicyTriggerContext,
 } from '../../types/generated'
 import { transformSystemData } from '../../utils/format'
 import { RawAzureBackupPolicyResource } from './data'
-import { BackupPolicy, TriggerContext } from './utils'
 
 const formatTrigger = (
   trigger: TriggerContext
@@ -17,22 +17,50 @@ const formatTrigger = (
   }
 
   return {
-    ...trigger,
+    objectType: trigger.objectType,
+    schedule: {
+      repeatingTimeIntervals: trigger.schedule?.repeatingTimeIntervals,
+      timeZone: trigger.schedule?.timeZone,
+    },
     taggingCriteria:
-      trigger?.taggingCriteria?.map(({ criteria, ...tc }) => ({
-        id: cuid(),
-        criteria:
-          criteria?.map(({ daysOfMonth, ...c }) => ({
-            id: cuid(),
-            daysOfMonth:
-              daysOfMonth?.map(d => ({
+      trigger?.taggingCriteria?.map(
+        ({ criteria, tagInfo, isDefault, taggingPriority }) => ({
+          id: cuid(),
+          tagInfo: {
+            id: tagInfo?.id,
+            tagName: tagInfo?.tagName,
+            eTag: tagInfo?.eTag,
+          },
+          isDefault,
+          taggingPriority,
+          criteria:
+            criteria?.map(
+              ({
+                daysOfMonth,
+                absoluteCriteria,
+                daysOfTheWeek,
+                monthsOfYear,
+                objectType,
+                scheduleTimes,
+                weeksOfTheMonth,
+              }) => ({
                 id: cuid(),
-                ...d,
-              })) || [],
-            ...c,
-          })) || [],
-        ...tc,
-      })) || [],
+                daysOfMonth:
+                  daysOfMonth?.map(d => ({
+                    id: cuid(),
+                    date: d.date,
+                    isLast: d.isLast,
+                  })) || [],
+                absoluteCriteria,
+                daysOfTheWeek,
+                monthsOfYear,
+                objectType,
+                scheduleTimes,
+                weeksOfTheMonth,
+              })
+            ) || [],
+        })
+      ) || [],
   }
 }
 
@@ -43,26 +71,81 @@ const formatProperties = (
     return {}
   }
 
-  const { policyRules, ...rest } = properties
+  const { policyRules, datasourceTypes, objectType } = properties
 
   return {
+    datasourceTypes,
+    objectType,
     policyRules:
-      policyRules?.map(({ trigger = {}, lifecycles = [], ...r }) => ({
-        id: cuid(),
-        trigger: formatTrigger(trigger),
-        lifecycles:
-          lifecycles?.map(({ targetDataStoreCopySettings, ...l }) => ({
-            id: cuid(),
-            ...l,
-            targetDataStoreCopySettings:
-              targetDataStoreCopySettings?.map(t => ({
+      policyRules?.map(
+        ({
+          trigger = {},
+          lifecycles = [],
+          backupParameters = {},
+          dataStore = {},
+          name,
+          objectType: policyRuleObjectType,
+          isDefault,
+        }) => ({
+          id: cuid(),
+          name,
+          objectType: policyRuleObjectType,
+          isDefault,
+          backupParameters: backupParameters
+            ? {
+                backupType: backupParameters.backupType,
+                objectType: backupParameters.objectType,
+              }
+            : {},
+          dataStore: dataStore
+            ? {
+                dataStoreType: dataStore.dataStoreType,
+                objectType: dataStore.objectType,
+              }
+            : {},
+          trigger: formatTrigger(trigger),
+          lifecycles:
+            lifecycles?.map(
+              ({
+                targetDataStoreCopySettings,
+                deleteAfter,
+                sourceDataStore,
+              }) => ({
                 id: cuid(),
-                ...t,
-              })) || [],
-          })) || [],
-        ...r,
-      })) || [],
-    ...rest,
+                deleteAfter: deleteAfter
+                  ? {
+                      duration: deleteAfter.duration,
+                      objectType: deleteAfter.objectType,
+                    }
+                  : {},
+                sourceDataStore: sourceDataStore
+                  ? {
+                      dataStoreType: sourceDataStore.dataStoreType,
+                      objectType: sourceDataStore.objectType,
+                    }
+                  : {},
+                targetDataStoreCopySettings:
+                  targetDataStoreCopySettings?.map(
+                    ({ copyAfter, dataStore: dataStoreCopySettings }) => ({
+                      id: cuid(),
+                      copyAfter: copyAfter
+                        ? {
+                            duration: copyAfter.duration,
+                            objectType: copyAfter.objectType,
+                          }
+                        : {},
+                      dataStore: dataStoreCopySettings
+                        ? {
+                            dataStoreType: dataStoreCopySettings.dataStoreType,
+                            objectType: dataStoreCopySettings.objectType,
+                          }
+                        : {},
+                    })
+                  ) || [],
+              })
+            ) || [],
+        })
+      ) || [],
   }
 }
